GenkitGo/surya-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
surya-ocr==0.6.7
pillow==10.4.0
pydantic==2.9.2
